% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{boxplotly}
\alias{boxplotly}
\title{Draw boxplot and save it as JSON and PDF files.}
\usage{
boxplotly(d, output_file = "", filetype = "pdf", ...)
}
\arguments{
\item{d}{A dataframe that contains the columns: gene_symbol, group, value.}

\item{output_file}{File path to save the output.}

\item{filetype}{The file format to save as (e.g., "pdf", "png").}

\item{...}{Additional arguments passed to the boxplot function.}
}
\value{
A ggplot object.
}
\description{
Draw boxplot and save it as JSON and PDF files.
}
