% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{query_data}
\alias{query_data}
\title{Query expression data and sample info from external files (expression.csv, sample_info.csv)}
\usage{
query_data(
  exp_file,
  sample_info_file,
  which_entrez_ids = NULL,
  which_gene_symbols = NULL,
  which_groups = NULL
)
}
\arguments{
\item{exp_file}{Path to the expression file.}

\item{sample_info_file}{Path to the sample info file.}

\item{which_entrez_ids}{Entrez IDs to filter the expression data.}

\item{which_gene_symbols}{Gene symbols to filter the expression data.}

\item{which_groups}{Groups to filter the sample info.}
}
\value{
A dataframe containing the expression data and sample info.
}
\description{
Query expression data and sample info from external files (expression.csv, sample_info.csv)
}
\examples{
# Two groups, two genes
d <- query_data(exp_file = "./examples/gene_expression.tsv", sample_info_file = "./examples/sample_info.tsv", which_entrez_ids = c("SL003951", "SL003970"), which_groups = c("Female_MECFS", "Female_Control"))

# One group, multiple genes
d <- query_data(exp_file = "./examples/gene_expression.tsv", sample_info_file = "./examples/sample_info.tsv", which_gene_symbols = c("TP53", "ERBB2"), which_groups = c("Female_MECFS"))

# Two groups, multiple genes
d <- query_data(exp_file = "./examples/gene_expression.tsv", sample_info_file = "./examples/sample_info.tsv", which_entrez_ids = c("SL003951", "SL003970", "SL003980"), which_groups = c("Female_MECFS", "Female_Control", "Male_MECFS"))
}
