% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-deg.R
\name{run_deg_with_deseq2}
\alias{run_deg_with_deseq2}
\title{\link{Not Ready} Run differential expression analysis using DESeq2}
\usage{
run_deg_with_deseq2(
  counts_file,
  sample_info_file,
  test_group_name,
  control_group_name,
  output_file
)
}
\arguments{
\item{counts_file}{Path to the counts data file: the columns are samples, the rows are genes; the first column is the gene/protein id (id), the second column is the gene symbol (gene_symbol), the remaining columns are the counts}

\item{sample_info_file}{Path to the sample information file: the file should have at least two columns, sample_id and group and the sample_id should be the same as the column names of the counts data file}

\item{test_group_name}{Name of the test group: the name of the test group should be one of the group in the sample information file}

\item{control_group_name}{Name of the control group: the name of the control group should be one of the group in the sample information file}

\item{output_file}{Path to the output file: the output file will be a tab-separated file with the following columns: id, gene_symbol, and the rest of the columns are the log-fold changes, t-statistics, p-values, and adjusted p-values}
}
\value{
None
}
\description{
\link{Not Ready} Run differential expression analysis using DESeq2
}
\examples{
run_deg_with_deseq2("counts.txt", "sample_info.txt", "test_group", "control_group", "output.txt")

}
